rename user settings to profile, move to evaluation app and show for all users
