+++ evap/contributor/templates/contributor_index.html
@@ -20,7 +20,7 @@
             <b>{% trans 'Delegates' %}</b><br />
             {% blocktrans %}Lecturers can assign delegates to help them with the preparation of the evaluation.{% endblocktrans %}
             {% if user.is_editor %}
-                <a href="{% url 'contributor:settings_edit' %}">{% blocktrans %}You can assign your own delegates on your settings page.{% endblocktrans %}</a><br />
+                <a href="{% url 'evaluation:profile_edit' %}">{% blocktrans %}You can assign your own delegates on your profile page.{% endblocktrans %}</a><br />
             {% endif %}
             {% if user.is_delegate %}
                 {% blocktrans %}Evaluations from lecturers who set you as a delegate are marked with a label below.{% endblocktrans %}<br />
