--- evap/contributor/templates/contributor_settings.html
@@ -1,100 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}{% trans 'Your EvaP' %} - {{ block.super }}{% endblock %}
-
-{% block content %}
-    {{ block.super }}
-    <h3>{% trans 'Settings' %}</h3>
-
-    <div class="row mb-3">
-        <div class="col-6">
-            <div class="card">
-                <div class="card-body">
-                    <h4 class="card-title mb-4">{% trans 'Personal information' %}</h4>
-                    <table class="table table-striped table-vertically-aligned">
-                        <tr>
-                            <td><strong>{% trans 'Title' %}</strong></td>
-                            <td>{{ form.instance.title|default_if_none:"" }}</td>
-                        </tr>
-                        <tr>
-                            <td><strong>{% trans 'First name' %}</strong></td>
-                            <td>{{ form.instance.first_name|default_if_none:"" }}</td>
-                        </tr>
-                        <tr>
-                            <td><strong>{% trans 'Last name' %}</strong></td>
-                            <td>{{ form.instance.last_name|default_if_none:"" }}</td>
-                        </tr>
-                        <tr>
-                            <td><strong>{% trans 'Email' %}</strong></td>
-                            <td>{{ form.instance.email|default_if_none:"" }}</td>
-                        </tr>
-                        <tr>
-                            <td><strong>{% trans 'CC all my emails to' %}</strong></td>
-                            <td>
-                                {% for cc_user in cc_users %}
-                                    {{ cc_user.full_name }}{% if not forloop.last %}, {% endif %}
-                                {% endfor %}
-                            </td>
-                        </tr>
-                        <tr>
-                            <td><strong>{% blocktrans %}You receive CC'ed emails for{% endblocktrans %}</strong></td>
-                            <td>
-                                {% for ccing_user in ccing_users %}
-                                    {{ ccing_user.full_name }}{% if not forloop.last %}, {% endif %}
-                                {% endfor %}
-                            </td>
-                        </tr>
-                    </table>
-                </div>
-                <div class="card-footer text-center">
-                    <button type="button" class="btn btn-light" onclick="changeRequestModalShow();" id="changeRequestModalButton">
-                        {% trans 'Request changes' %}
-                    </button>
-                </div>
-            </div>
-        </div>
-        <div class="col-6">
-            <form id="settings-form" method="POST" class="form-horizontal multiselect-form settings-form" enctype="multipart/form-data">
-                {% csrf_token %}
-                {% include 'bootstrap_form_errors.html' with errors=form.non_field_errors %}
-                <div class="card">
-                    <div class="card-body">
-                        <h4 class="card-title mb-4">{% trans 'Delegates' %}</h4>
-                        <p class="mb-3">
-                            {% trans 'Your delegates will receive all your emails in CC and will be able to prepare evaluations and see results on your behalf.' %}
-                        </p>
-                        <table class="table table-vertically-aligned">
-                            <tr>
-                                <td>
-                                <strong>{% trans 'Your delegates' %}</strong><br />
-                                {% include 'bootstrap_form_errors.html' with errors=form.delegates.errors %}
-                                {{ form.delegates }}
-                                </td>
-                            </tr>
-                            {% if delegate_of %}
-                                <tr>
-                                    <td>
-                                        <strong>{% trans 'You are a delegate of' %}</strong><br />
-                                        {% for delegate in delegate_of %}
-                                            {{ delegate.full_name }}{% if not forloop.last %}, {% endif %}
-                                        {% endfor %}
-                                    </td>
-                                </tr>
-                            {% endif %}
-                        </table>
-                    </div>
-                    <div class="card-footer text-center">
-                        <button type="submit" class="btn btn-primary">{% trans 'Save delegates' %}</button>
-                    </div>
-                </div>
-            </form>
-        </div>
-    </div>
-{% endblock %}
-
-{% block modals %}
-    {{ block.super }}
-    {% trans 'Request user setting changes' as title %}
-    {% trans 'Please tell us which of your settings we should change.' as teaser %}
-    {% include 'contact_modal.html' with modal_id='changeRequestModal' user=request.user title=title teaser=teaser %}
-{% endblock %}
