+++ evap/contributor/urls.py
@@ -7,7 +7,6 @@ app_name = "contributor"
 
 urlpatterns = [
     path("", views.index, name="index"),
-    path("settings", views.settings_edit, name="settings_edit"),
     path("export", views.export, name="export"),
     path("evaluation/<int:evaluation_id>", views.evaluation_view, name="evaluation_view"),
     path("evaluation/<int:evaluation_id>/edit", views.evaluation_edit, name="evaluation_edit"),
