+++ evap/evaluation/tests/test_forms.py
@@ -2,8 +2,9 @@ from django.test import TestCase
 
 from model_bakery import baker
 
+from evap.evaluation.forms import DelegatesForm, NewKeyForm
 from evap.evaluation.models import UserProfile
-from evap.evaluation.forms import NewKeyForm
+from evap.evaluation.tests.tools import get_form_data_from_instance
 
 
 class TestNewKeyForm(TestCase):
@@ -20,3 +21,25 @@ class TestNewKeyForm(TestCase):
         form = NewKeyForm(data)
         self.assertFalse(form.is_valid())
         self.assertIn("Inactive users cannot request login keys.", form.errors["email"])
+
+
+class UserFormTests(TestCase):
+
+    def test_settings_form(self):
+        """
+            Tests whether the settings form can be submitted without errors
+        """
+        user = baker.make(UserProfile, email="testuser@institution.example.com")
+        delegate = baker.make(UserProfile, email="delegate@institution.example.com")
+
+        self.assertFalse(user.delegates.filter(email="delegate@institution.example.com").exists())
+
+        form_data = get_form_data_from_instance(DelegatesForm, user)
+        form_data["delegates"] = [delegate.pk]  # add delegate
+
+        form = DelegatesForm(form_data, instance=user)
+        self.assertTrue(form.is_valid())
+        form.save()
+
+        user = UserProfile.objects.get(email="testuser@institution.example.com")
+        self.assertTrue(user.delegates.filter(email="delegate@institution.example.com").exists())
